analysis/requirements.txt
matplotlib
psycopg2-binary
scipy
numpy
PyGObject
pycairo
